+++ energy_orchestrator/app/ml/__init__.py
@@ -1,4 +1,9 @@
-from .heating_features import build_heating_feature_dataset
+from .heating_features import (
+    build_heating_feature_dataset,
+    compute_scenario_historical_features,
+    get_actual_vs_predicted_data,
+    validate_prediction_start_time,
+)
 from .heating_demand_model import (
     HeatingDemandModel,
     train_heating_demand_model,
@@ -9,6 +14,9 @@ from .heating_demand_model import (
 
 __all__ = [
     "build_heating_feature_dataset",
+    "compute_scenario_historical_features",
+    "get_actual_vs_predicted_data",
+    "validate_prediction_start_time",
     "HeatingDemandModel",
     "train_heating_demand_model",
     "load_heating_demand_model",